LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-authentication-basic-cors-not-enabled-expected.txt

Blocked access to external URL http://www1.localhost:8800/XMLHttpRequest/resources/auth1/auth.py
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load http://www1.localhost:8800/XMLHttpRequest/resources/auth1/auth.py due to access control checks.

PASS XMLHttpRequest: send() - "Basic" authenticated CORS requests with user name and password passed to open() (asserts failure) 
